Refactor tests for layouter
